Use ErrorCodes directly as keys in unordered maps

+++ arangod/Cluster/ClusterInfo.cpp
@@ -4550,7 +4550,7 @@ Result ClusterInfo::dropIndexCoordinatorInner(
   AgencyCommResult result = ac.sendTransactionWithFailover(trx, 0.0);
 
   if (!result.successful()) {
-    if (result.httpCode() == (int)arangodb::rest::ResponseCode::PRECONDITION_FAILED) {
+    if (result.httpCode() == arangodb::rest::ResponseCode::PRECONDITION_FAILED) {
       // Retry loop is outside!
       return Result(TRI_ERROR_HTTP_PRECONDITION_FAILED);
     }
